+++ R/project.R
@@ -43,9 +43,9 @@ NULL
 #'   are appended to the previous ones. Only relevant if `object` is a
 #'   `MizerSim` object. Default = TRUE.
 #' @param progress_bar Either a boolean value to determine whether a progress
-#'   bar should be shown in the console of a shiny progress object to implement 
-#'   a progress bar in a shiny app
-#' @param ... Currently unused.
+#'   bar should be shown in the console, or a shiny progress object to implement 
+#'   a progress bar in a shiny app.
+#' @param ... Other arguments will be passed to rate functions.
 #' 
 #' @note The `effort` argument specifies the level of fishing effort during
 #' the simulation. If it is not supplied, the initial effort stored in the params
@@ -61,7 +61,7 @@ NULL
 #' constant through time but each gear may have a different fishing effort. 
 #' \item A numerical array with dimensions time x gear. This specifies the
 #' fishing effort of each gear at each time step.  The first dimension, time,
-#' must be named numerically and contiguously. The second dimension of the array
+#' must be named numerically and increasing. The second dimension of the array
 #' must be named and the names must correspond to the gear names in the
 #' `MizerParams` argument. The value for the effort for a particular time
 #' is used during the interval from that time to the next time in the array.
@@ -82,8 +82,6 @@ NULL
 #' @return An object of class \linkS4class{MizerSim}.
 #' 
 #' @export
-#' @seealso [MizerParams()], [summary_functions()],
-#' [indicator_functions()] and [plotting_functions()]
 #' @examples
 #' \dontrun{
 #' # Data set with different fishing gears
@@ -110,6 +108,8 @@ project <- function(object, effort,
                     append = TRUE,
                     progress_bar = TRUE, ...) {
     validObject(object)
+    
+    # Set and check initial values ----
     if (is(object, "MizerSim")) {
         params <- object@params
         no_t <- dim(object@n)[[1]]
@@ -119,11 +119,13 @@ project <- function(object, effort,
         initial_n_pp[] <- object@n_pp[no_t, ]
         initial_n_other <- object@n_other[no_t, ]
         t_start <- as.numeric(dimnames(object@n)[[1]][[no_t]])
-    } else {
+    } else if (is(object, "MizerParams")) {
         params <- object
         if (missing(initial_n))       initial_n <- params@initial_n
         if (missing(initial_n_pp)) initial_n_pp <- params@initial_n_pp
         initial_n_other <- params@initial_n_other
+    } else {
+        stop("The `object` argument must be either a MizerParams or a MizerSim object.")
     }
     params@initial_n[] <- initial_n
     params@initial_n_pp[] <- initial_n_pp
@@ -148,10 +150,10 @@ project <- function(object, effort,
         }
     }
     
-    # Create effort array ----
+    # Set effort array ----
     if (missing(effort)) effort <- params@initial_effort
     # Do we need to create an effort array?
-    if (is.null(dim(effort))) {
+    if (is.null(dim(effort))) { # effort is a vector
         no_gears <- dim(params@catchability)[1]
         if ((length(effort) > 1) & (length(effort) != no_gears)) {
             stop("Effort vector must be the same length as the number of fishing gears\n")
@@ -159,7 +161,7 @@ project <- function(object, effort,
         # If more than 1 gear need to check that gear names match
         gear_names <- dimnames(params@catchability)[[1]]
         effort_gear_names <- names(effort)
-        if (length(effort) == 1 & is.null(effort_gear_names)) {
+        if (length(effort) == 1 && is.null(effort_gear_names)) {
             effort_gear_names <- gear_names
         }
         if (!all(gear_names %in% effort_gear_names)) {
@@ -168,10 +170,8 @@ project <- function(object, effort,
                  ") do not match those in the effort vector.")
         }
         # Set up the effort array transposed so we can use the recycling rules
-        time_dimnames <- seq(from = t_start, 
-                             to = t_start + t_max, 
-                             by = dt)
-        effort <- t(array(effort, dim = c(no_gears, length(time_dimnames)), 
+        time_dimnames <- c(t_start, t_start + t_max)
+        effort <- t(array(effort, dim = c(no_gears, 2), 
                           dimnames = list(gear = effort_gear_names, 
                                           time = time_dimnames)))
     }
@@ -194,9 +194,6 @@ project <- function(object, effort,
     # Sort effort array to match order in MizerParams
     effort <- effort[, gear_names, drop = FALSE]
     
-    # Blow up time dimension of effort array
-    # i.e. effort might have been passed in using time steps of 1, but actual 
-    # dt = 0.1, so need to blow up
     if (is.null(dimnames(effort)[[1]])) {
         stop("The time dimname of the effort argument must be numeric.")
     }
@@ -207,156 +204,100 @@ project <- function(object, effort,
     if (is.unsorted(time_effort)) {
         stop("The time dimname of the effort argument should be increasing.")
     }
-    t_end <- time_effort[length(time_effort)]
-    # Blow up effort so that rows are dt spaced
-    time_effort_dt <- seq(from = time_effort[1], to = t_end, by = dt)
-    effort_dt <- t(array(NA, dim = c(length(time_effort_dt), dim(effort)[2]), 
-                         dimnames = list(time = time_effort_dt,
-                                         dimnames(effort)[[2]])))
-    for (i in 1:(length(time_effort) - 1)) {
-        effort_dt[,time_effort_dt >= time_effort[i]] <- effort[i,]
-    }
-    effort_dt <- t(effort_dt)
+    t_start <- time_effort[[1]]
+    t_end <- time_effort[[length(time_effort)]]
+    t_max <- t_end - t_start
     
     # Make the MizerSim object with the right size ----
-    # We only save every t_save steps
+    # We only save every t_save years
+    # no point running a simulation with no saved results
+    if (t_max < t_save) {
+        t_save <- t_max
+    }
     # Divisibility test needs to be careful about machine rounding errors,
     # see https://github.com/sizespectrum/mizer/pull/2
     if ((t_save < dt) || !isTRUE(all.equal((t_save - round(t_save / dt) * dt), 0)))
         stop("t_save must be a positive multiple of dt")
-    t_skip <- round(t_save/dt)
-    t_dimnames_index <- seq(1, to = length(time_effort_dt), by = t_skip)
-    t_dimnames <- time_effort_dt[t_dimnames_index]
-    sim <- MizerSim(params, t_dimnames = t_dimnames) 
-    # Fill up the effort array
-    sim@effort[] <- effort_dt[t_dimnames_index,]
+    skip <- round(t_save/dt)
+    t_dimnames <- seq(t_start, t_end, by = t_save)
+    sim <- MizerSim(params, t_dimnames = t_dimnames)
     
     ## Initialise ----
-    # Set initial population
-    sim@n[1, , ] <- initial_n 
-    sim@n_pp[1, ] <- initial_n_pp
-    sim@n_other[1, ] <- initial_n_other
-    
-    # Handy things
-    no_sp <- nrow(sim@params@species_params) # number of species
-    no_w <- length(sim@params@w) # number of fish size bins
-    idx <- 2:no_w
-    resource_dynamics_fn <- get(sim@params@resource_dynamics)
-    other_dynamics_fns <- lapply(sim@params@other_dynamics, get)
-    rates_fns <- lapply(sim@params@rates_funcs, get)
-    # Hacky shortcut to access the correct element of a 2D array using 1D notation
-    # This references the egg size bracket for all species, so for example
-    # n[w_minidx_array_ref] = n[,w_min_idx]
-    w_min_idx_array_ref <- (sim@params@w_min_idx - 1) * no_sp + (1:no_sp)
-    
-    # Matrices for solver
-    a <- matrix(0, nrow = no_sp, ncol = no_w)
-    b <- matrix(0, nrow = no_sp, ncol = no_w)
-    S <- matrix(0, nrow = no_sp, ncol = no_w)
-    
-    # initialise n n_pp and n_other
-    # We want the first time step only but cannot use drop as there may only 
-    # be a single species
-    n <- initial_n
-    n_pp <- initial_n_pp
-    n_other <- initial_n_other
+    # get functions
+    resource_dynamics_fn <- get(params@resource_dynamics)
+    other_dynamics_fns <- lapply(params@other_dynamics, get)
+    rates_fns <- lapply(params@rates_funcs, get)
     
     # Set up progress bar
     if (progress_bar == TRUE) {
         pb <- progress::progress_bar$new(
             format = "[:bar] :percent ETA: :eta",
-            total = length(t_dimnames_index), width = 60)
+            total = length(t_dimnames), width = 60)
+        pb$tick(0)
     }
     if (is(progress_bar, "Progress")) {
         # We have been passed a shiny progress object
         progress_bar$set(message = "Running simulation", value = 0)
-        proginc <- 1/length(t_dimnames_index)
+        proginc <- 1/length(t_dimnames)
     }
     
+    t <- t_start  # keep track of time
+    i_save_time <- 2 # index into which to save the results
+    
+    # set up variables for keeping track of effort changes
+    i_effort <- 1
+    current_effort <- effort[1, ]
+    t_next_effort <- time_effort[[2]] - 1e-8  # to avoid rounding problems
+    
+    # Set initial population and effort
+    n_list <- list(n = initial_n, n_pp = initial_n_pp, n_other = initial_n_other)
+    sim@n[1, , ] <- initial_n 
+    sim@n_pp[1, ] <- initial_n_pp
+    sim@n_other[1, ] <- initial_n_other
+    sim@effort[1, ] <- current_effort
+    
     ## Loop over time ----
-    t <- 0  # keep track of time
-    t_steps <- dim(effort_dt)[1] - 1
-    for (i_time in 1:t_steps) {
-        r <- rates_fns$Rates(
-            params, n = n, n_pp = n_pp, n_other = n_other,
-            t = t, effort = effort_dt[i_time,], rates_fns = rates_fns)
+    for (i in 2:length(t_dimnames)) {
         
-        # Update time
-        t <- t + dt
-        # Update other components
-        n_other_current <- n_other  # So that the resource dynamics can still 
-                                    # use the current value
-        for (res in other_names) {
-            n_other[[res]] <-
-                other_dynamics_fns[[res]](
-                    params,
-                    n = n,
-                    n_pp = n_pp,
-                    n_other = n_other_current,
-                    rates = r,
-                    t = t,
-                    dt = dt
-                )
+        n_list <- project_simple(
+            params, n = n_list$n, n_pp = n_list$n_pp, n_other = n_list$n_other,
+            t = t, dt = dt, steps = skip, effort = current_effort,
+            resource_dynamics_fn = resource_dynamics_fn,
+            other_dynamics_fns = other_dynamics_fns,
+            rates_fns = rates_fns)
+        t <- t + t_save
+        
+        # Advance progress bar
+        if (is(progress_bar, "Progress")) {
+            progress_bar$inc(amount = proginc)
+        }
+        if (progress_bar == TRUE) {
+            pb$tick()
         }
         
-        # Update resource
-        n_pp <- resource_dynamics_fn(params, n = n, n_pp = n_pp,
-                                     n_other = n_other_current, rates = r,
-                                     t = t, dt = dt)
-        
-        # Iterate species one time step forward:
-        # See Ken's PDF
-        # a_{ij} = - g_i(w_{j-1}) / dw_j dt
-        a[, idx] <- sweep(-r$e_growth[, idx - 1, drop = FALSE] * dt, 2,
-                          sim@params@dw[idx], "/")
-        # b_{ij} = 1 + g_i(w_j) / dw_j dt + \mu_i(w_j) dt
-        b[, idx] <- 1 + sweep(r$e_growth[, idx, drop = FALSE] * dt, 2, 
-                              sim@params@dw[idx], "/") +
-                        r$mort[, idx, drop = FALSE] * dt
-        # S_{ij} <- N_i(w_j)
-        S[,idx] <- n[, idx, drop = FALSE]
-        # Boundary condition upstream end (reproduction)
-        b[w_min_idx_array_ref] <- 1 + r$e_growth[w_min_idx_array_ref] * dt /
-                                        sim@params@dw[sim@params@w_min_idx] +
-                                    r$mort[w_min_idx_array_ref] * dt
-        # Update first size group of n
-        n[w_min_idx_array_ref] <-
-            (n[w_min_idx_array_ref] + r$rdd * dt / 
-                 sim@params@dw[sim@params@w_min_idx]) /
-            b[w_min_idx_array_ref]
-        # Update n
-        # for (i in 1:no_sp) # number of species assumed small, so no need to 
-        #                      vectorize this loop over species
-        #     for (j in (sim@params@w_min_idx[i]+1):no_w)
-        #         n[i,j] <- (S[i,j] - A[i,j]*n[i,j-1]) / B[i,j]
-        # This is implemented via Rcpp
-        n <- inner_project_loop(no_sp = no_sp, no_w = no_w, n = n,
-                                A = a, B = b, S = S,
-                                w_min_idx = sim@params@w_min_idx)
-        
-        # Store results only every t_step steps.
-        store <- t_dimnames_index %in% (i_time + 1)
-        if (any(store)) {
-            # Advance progress bar
-            if (is(progress_bar, "Progress")) {
-                progress_bar$inc(amount = proginc)
-            }
-            if (progress_bar == TRUE) {
-                pb$tick()
-            }
-            # Store result
-            t_idx <- which(store)
-            sim@n[t_idx, , ] <- n
-            sim@n_pp[t_idx, ] <- n_pp
-            sim@n_other[t_idx, ] <- n_other
+        # Do we need new effort?
+        if (t >= t_next_effort) {
+            i_effort <- i_effort + 1
+            current_effort <- effort[i_effort, ]
+            # We have to be careful when we reach the final time
+            t_next_effort <- ifelse(i_effort < length(time_effort),
+                                    time_effort[[i_effort + 1]] - 1e-8,
+                                    t_max + 100)
         }
+        
+        # Store result
+        sim@n[i, , ] <- n_list$n
+        sim@n_pp[i, ] <- n_list$n_pp
+        sim@n_other[i, ] <- n_list$n_other
+        sim@effort[i, ] <- current_effort
     }
+    
+    # append to previous simulation ----
     if (is(object, "MizerSim") && append) {
-        # append to previous simulation ----
         no_t_old <- dim(object@n)[1]
         no_t <- length(t_dimnames)
         new_t_dimnames <- c(as.numeric(dimnames(object@n)[[1]]),
-                            t_dimnames[2:length(t_dimnames)])
+                            t_dimnames[2:no_t])
         new_sim <- MizerSim(params, t_dimnames = new_t_dimnames)
         old_indices <- 1:no_t_old
         new_indices <- seq(from = no_t_old + 1, length.out = no_t - 1)
@@ -373,6 +314,137 @@ project <- function(object, effort,
     return(sim)
 }
 
+#' Project abundances by a given number of time steps into the future
+#' 
+#' This is an internal function used by the user-facing `project()` function.
+#' It is of potential interest only to mizer extension authors.
+#' 
+#' The function does not check its arguments because it is mean to be as fast
+#' as possible to allow it to be used in a loop. For example it is called in
+#' `project()` once for every saved value. The function also does not save its
+#' intermediate results but only returns the result at time `t + dt * steps`.
+#' During this time it uses the constant fishing effort `effort`.
+#' 
+#' The functional arguments can be calculated from slots in the `params` object
+#' with
+#' ```
+#' resource_dynamics_fn <- get(params@resource_dynamics)
+#' other_dynamics_fns <- lapply(params@other_dynamics, get)
+#' rates_fns <- lapply(params@rates_funcs, get)
+#' ```
+#' The reason the function does not do that itself is to shave 20 microseconds
+#' of its running time, which pays when the function is called hundreds of
+#' times in a row.
+#' 
+#' This function is also used in `steady()`. In between calls to 
+#' `project_simple()` the `steady()` function checks whether the values are
+#' still changing significantly, so that it can stop when a steady state has
+#' been approached. Mizer extension packages might have a similar need to run
+#' a simulation repeatedly for short periods to run some other code in
+#' between. Because this code may want to use the values of the rates at the
+#' final time step, these too are included in the returned list.
+#' 
+#' @param params A MizerParams object.
+#' @param n An array (species x size) with the number density at start of
+#'   simulation.
+#' @param n_pp A vector (size) with the resource number density at start of
+#'   simulation.
+#' @param n_other A named list with the abundances of other components at start
+#'   of simulation.
+#' @param t Time at the start of the simulation.
+#' @param dt Size of time step.
+#' @param steps The number of time steps by which to project.
+#' @param effort The fishing effort to be used throughout the simulation. This
+#'   must be a vector or list with one named entry per fishing gear.
+#' @param resource_dynamics_fn The function for the resource
+#'   dynamics. See Details.
+#' @param other_dynamics_fns List with the functions for the
+#'   dynamics of the other components. See Details.
+#' @param rates_fns List with the functions for calculating
+#'   the rates. See Details.
+#' @param ... Other arguments that are passed on to the rate functions.
+#' @return List with the final values of `n`, `n_pp` and `n_other`, `rates`.
+#' 
+#' @md
+#' @export
+project_simple <- function(params, n, n_pp, n_other, t, dt, steps, 
+                           effort, resource_dynamics_fn, other_dynamics_fns,
+                           rates_fns, ...) {    
+    # Handy things
+    no_sp <- nrow(params@species_params) # number of species
+    no_w <- length(params@w) # number of fish size bins
+    idx <- 2:no_w
+    # Hacky shortcut to access the correct element of a 2D array using 1D notation
+    # This references the egg size bracket for all species, so for example
+    # n[w_minidx_array_ref] = n[,w_min_idx]
+    w_min_idx_array_ref <- (params@w_min_idx - 1) * no_sp + (1:no_sp)
+    # Matrices for solver
+    a <- matrix(0, nrow = no_sp, ncol = no_w)
+    b <- matrix(0, nrow = no_sp, ncol = no_w)
+    S <- matrix(0, nrow = no_sp, ncol = no_w)
+    
+    for (i_time in 1:steps) {
+        r <- rates_fns$Rates(
+            params, n = n, n_pp = n_pp, n_other = n_other,
+            t = t, effort = effort, rates_fns = rates_fns, ...)
+        
+        # Update time
+        t <- t + dt
+        
+        # Update other components
+        n_other_current <- n_other  # So that the resource dynamics can still 
+        # use the current value
+        for (component in names(params@other_dynamics)) {
+            n_other[[component]] <-
+                other_dynamics_fns[[component]](
+                    params,
+                    n = n,
+                    n_pp = n_pp,
+                    n_other = n_other_current,
+                    rates = r,
+                    t = t,
+                    dt = dt,
+                    component = component,
+                    ...
+                )
+        }
+        
+        # Update resource
+        n_pp <- resource_dynamics_fn(params, n = n, n_pp = n_pp,
+                                     n_other = n_other_current, rates = r,
+                                     t = t, dt = dt, ...)
+        
+        # Iterate species one time step forward:
+        # a_{ij} = - g_i(w_{j-1}) / dw_j dt
+        a[, idx] <- sweep(-r$e_growth[, idx - 1, drop = FALSE] * dt, 2,
+                          params@dw[idx], "/")
+        # b_{ij} = 1 + g_i(w_j) / dw_j dt + \mu_i(w_j) dt
+        b[, idx] <- 1 + sweep(r$e_growth[, idx, drop = FALSE] * dt, 2, 
+                              params@dw[idx], "/") +
+            r$mort[, idx, drop = FALSE] * dt
+        # S_{ij} <- N_i(w_j)
+        S[,idx] <- n[, idx, drop = FALSE]
+        # Boundary condition upstream end (reproduction)
+        b[w_min_idx_array_ref] <- 1 + r$e_growth[w_min_idx_array_ref] * dt /
+            params@dw[params@w_min_idx] +
+            r$mort[w_min_idx_array_ref] * dt
+        # Update first size group of n
+        n[w_min_idx_array_ref] <-
+            (n[w_min_idx_array_ref] + r$rdd * dt / 
+                 params@dw[params@w_min_idx]) /
+            b[w_min_idx_array_ref]
+        # Update n
+        # for (i in 1:no_sp) # number of species assumed small, so no need to 
+        #                      vectorize this loop over species
+        #     for (j in (params@w_min_idx[i]+1):no_w)
+        #         n[i,j] <- (S[i,j] - A[i,j]*n[i,j-1]) / B[i,j]
+        # This is implemented via Rcpp
+        n <- inner_project_loop(no_sp = no_sp, no_w = no_w, n = n,
+                                A = a, B = b, S = S,
+                                w_min_idx = params@w_min_idx)
+    }
+    return(list(n = n, n_pp = n_pp, n_other = n_other, rates = r))
+}
 
 #' Calculate initial population abundances for the community populations
 #'